Use pytorch make_tensor and add licenses
