+++ python_tests/pytest_opinfos.py
@@ -1,3 +1,8 @@
+# SPDX-FileCopyrightText: Copyright (c) 2023-present NVIDIA CORPORATION & AFFILIATES.
+# All rights reserved.
+# SPDX-License-Identifier: BSD-3-Clause
+# Owner(s): ["module: nvfuser"]
+
 import torch
 import jax
 from pytest_core import (
